% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{asym_trimmed_perm_distribution}
\alias{asym_trimmed_perm_distribution}
\title{Permutation distribution for asymmetrically trimmed statistics}
\usage{
asym_trimmed_perm_distribution(x, y, type, randomization = FALSE, n.rep = NULL)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{type}{specifies the skewness selector statistic used for trimming, must be in
\code{"Q2", "SK2"} and \code{"SK5"}. Default is \code{"Q2"}.}

\item{randomization}{logical value indicating whether the p-value should be computed from a permutation (\code{FALSE}) or a randomization (\code{TRUE}) distribution.}

\item{n.rep}{an integer value specifying the number of random splits used to
calculate the randomization distribution if \code{method = "randomization"},
ignored if \code{method = "permutation"} or \code{method = "asymptotic"};
default is \code{n.rep = 10000}.}
}
\value{
Vector with permutation distribution.
}
\description{
\code{asym_trimmed_perm_distribution} calculates the permutation distribution for the asymmetrically
trimmed statistics from \code{asym_trimmed_test}.
}
