% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{asym_trimmed_t}
\alias{asym_trimmed_t}
\title{Test statistic for the asymmetrically trimmed test}
\usage{
asym_trimmed_t(x, y, type, delta = 0, na.rm = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{type}{specifies the skewness selector statistic used for trimming, must be in
\code{"Q2", "SK2"} and \code{"SK5"}. Default is \code{"Q2"}.}

\item{delta}{a numeric value indicating the true difference in the location or scale parameter, depending
on whether the test should be performed for a difference in location or in scale, default is
\code{delta = 0} for location test and \code{delta = 1} for scale test.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}
}
\value{
A list containing the following components:
\item{statistic}{the value of the test statistic.}
\item{estimates}{the asymmetrically trimmed means for both samples.}
\item{df}{the degrees of freedom for the test statistic.}
}
\description{
\code{asym_trimmed_t} calculates the test statistic of the asymmetrically trimmed test with a given skewness selector statistic
}
\references{
\insertRef{YueDix73appr}{robTests}

\insertRef{Yue74trim}{robTests}
}
