% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymTrimmedTest.R
\name{asym_trimmed_test}
\alias{asym_trimmed_test}
\title{Two-sample location tests based on asymmetrically trimmed means}
\usage{
asym_trimmed_test(x, y, type = c("Q2", "SK2", "SK5"),
  alternative = c("two.sided", "greater", "less"), delta = 0,
  method = c("asymptotic", "exact", "sampled"), n.rep = 10000,
  na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector of observations.}

\item{y}{a numeric vector of observations.}

\item{type}{specifies the skewness selector statistic used for trimming, must be in
\code{"Q2", "SK2"} and \code{"SK5"}. Default is \code{"Q2"}.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), 
"\code{greater}" or "\code{less}".}

\item{delta}{a numeric value indicating the true difference in the location parameter, default is \code{delta = 0}.}

\item{method}{a character string specifying the test method used, \code{"asymptotic"} for an asymptotic test based on the 
normal distribution, \code{"exact"} for an exact and \code{"sampled"} for a randomized permutation test. The exact 
permutation test uses all data permutations while the randomized test draws \code{n.rep} random permutations
with replacement.}

\item{n.rep}{an integer value specifying the number of random permutations used to calculate 
the permutation distribution if \code{method = "sampled"}, 
ignored if \code{method = "exact"} or \code{method = "asymptotic"}. Default is \code{n.rep = 10000}.}

\item{na.rm}{a logical value indicating whether NA values in \code{x} and \code{y} should be stripped before the computation proceeds.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the estimated sample means of \code{x} and \code{y}.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of trimmed t-test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
\code{asym_trimmed_test()} performs a two-sample location test by using
asymmetrically trimmed means based on different skewness-
selector statistics for both samples.
}
\references{
\insertRef{SmyPhi10perm}{robTests}
\insertRef{ReeSta04robu}{robTests}
}
