% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleEstimators.R
\name{asym_win_var}
\alias{asym_win_var}
\title{Asymmetrically trimmed variance}
\usage{
asym_win_var(x, type = c("Q2", "SK2", "SK5"), na.rm = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{type}{specifies the skewness selector statistic used for trimming, must be in
\code{"Q2", "SK2"} and \code{"SK5"}. Default is \code{"Q2"}.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}
}
\value{
A list containing the following items:
\item{var}{The asymmetrically trimmed variance, and}
\item{h}{the degrees of freedom used for tests based on  asymmetrically trimmed means and the
asymmetrically winsorized variance.}
}
\description{
\code{asym_win_var} calculates the asymmetrically trimmed variance using different
skewness selector statistics.
}
\examples{
set.seed(108)
x <- rnorm(10)
asym_win_var(x, type = "SK5")


}
