% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{calc_perm_p_value}
\alias{calc_perm_p_value}
\title{Calculation of permutation p-value}
\usage{
calc_perm_p_value(statistic, distribution, m, n, sampled, n.rep, alternative)
}
\arguments{
\item{statistic}{observed value of the test statistic.}

\item{distribution}{numeric vector with permutation distribution.}

\item{m}{integer value giving size of first sample.}

\item{n}{integer value giving size of second sample.}

\item{sampled}{logical value indicating whether all splits should be
considered or only a random sample, where the observations
are drawn with replacement.}

\item{n.rep}{an integer value specifying the number of random splits used to
calculate the permutation distribution if \code{method = "sampled"},
ignored if \code{method = "exact"} or \code{method = "asymptotic"};
default is \code{n.rep = 10000}.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}
}
\value{
p.value for the specified alternative.
}
\description{
\code{calc_perm_p_value} calculates the permutation p-value following Smyth & Phipson (2010).
}
\references{
\insertRef{SmyPhi10perm}{robTests}
}
