% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationEstimators.R
\name{hodges_lehmann_2sample}
\alias{hodges_lehmann_2sample}
\title{Two-sample Hodges-Lehmann estimator}
\usage{
hodges_lehmann_2sample(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{y}{numeric vector of observations.}

\item{na.rm}{a logical value indicating whether NA values in \code{x} and \code{y} 
should be stripped before the computation proceeds.}
}
\value{
The two-sample Hodges-Lehmann estimator.
}
\description{
\code{hodges_lehmann_2sample} calculates the two-sample Hodges-Lehmann estimator for the location difference
of two samples x and y.
}
\examples{
x <- rnorm(10); y <- rnorm(10)
hodges_lehmann_2sample(x, y)

}
\references{
\insertRef{HodLeh63esti}{robTests}
}
