% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{huber_2sample}
\alias{huber_2sample}
\title{Simultaneous Huber-M-estimates of scale and location}
\usage{
huber_2sample(x, y, k)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{k}{tuning parameter(s) for the respective kernel function,
defaults to parameters implemented in \code{\link[robustbase:lmrob.control]{.Mpsi.tuning.default(psi)}} in the package \pkg{robustbase}.}
}
\value{
Named list containing the following objects
        \item{mu.x}{Location estimate of x}
        \item{mu.y}{Location estimate of y}
        \item{s}{Scale estimate for the joined sample}
}
\description{
Calculates M-estimates of location and the joined scale of two samples
}
