% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberTest.R
\name{huber_test}
\alias{huber_test}
\title{Two-sample location tests based on Huber's M-estimator}
\usage{
huber_test(x, y, delta = 0, k = 1.8, alternative = c("two.sided",
  "greater", "less"))
}
\arguments{
\item{x}{a numeric vector of observations.}

\item{y}{a numeric vector of observations.}

\item{delta}{a numeric value indicating the true difference in the location parameter, default is \code{delta = 0}.}

\item{k}{Tuning parameter for Huber's M-estimator.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default),
"\code{greater}" or "\code{less}".}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{parameter}{the degrees of freedom for the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the Huber M-estimates of \code{x} and \code{y}.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
Performs a 2-sample location test based on Huber's M-estimator
}
\details{
The test is introduced in the context of the hybrid tests in Weichert & Hothorn (2002).
}
\examples{
x <- rnorm(20); y <- rnorm(20)
huber_test(x, y)

}
\references{
\insertRef{WeiHot02robu}{robTests}
}
\seealso{
\code{\link[robTests]{min_tc_test}}
\code{\link[robTests]{min_t_test}}
}
