% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridTest.R
\name{hybrid_test}
\alias{hybrid_test}
\title{Hybrid permutation test for location differences}
\usage{
hybrid_test(
  x,
  y,
  type = c("min1", "min2", "min3"),
  alternative = c("two.sided", "greater", "less"),
  delta = 0,
  k = 1.8,
  na.rm = FALSE,
  n.rep = 1000,
  var.test = FALSE
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{type}{a character string specifying the type of hybrid test used, must be one of \code{"min1"}, \code{"min2"} and
\code{"min3"}. See details for information on the three types.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{delta}{a numeric value indicating the true difference in the location or scale parameter, depending
on whether the test should be performed for a difference in location or in scale, default is
\code{delta = 0} for location test and \code{delta = 1} for scale test.}

\item{k}{tuning parameter for Huber's M-estimator, default is \code{k = 1.8},
only needed if \code{type \%in\% c("min2", "min3")}.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}

\item{n.rep}{an integer value specifying the number of random permutations used to calculate
the permutation distribution of the minimum p-value;
default is \code{n.rep = 1000}.}

\item{var.test}{logical value testing whether the samples should be compared for a difference in scale.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the minimum p-value of the test statistics used.}
\item{p.value}{the p-value for the test.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
\code{hybrid_test()} performs a hybrid test based on the p-values of different test statistics
}
\details{
The tests implemented here are introduced in Weichert & Hothorn (2002). They are based on
the minimum p-values from different test statistics. The test statistics used are specified in the
\code{type}-argument. We have three different types:
\describe{
\item{min1}{test based on the t-statistic and the 10\% and 20\% trimmed t-statistics}
\item{min2}{test based on the t-statistic and a test statistic based on Huber's M-estimator}
\item{min3}{test based on the t-statistic, the 20\% trimmed t-statistic and the test based on Huber's M-estimator}
}

The test statistic is the minumum p-value of the different test statistics.
The permutation distribution of the minimum p-value is achieved using the
permutation principle according to Efron & Tibshirani (1998).
}
\examples{
x <- rnorm(10); y <- rnorm(10)
hybrid_test(x, y, type = "min1")

}
\references{
\insertRef{WeiHot02robu}{robTests}

\insertRef{EfrTib98intr}{robTests}
}
\seealso{
\code{\link[stats]{t.test}}
 \code{\link[robTests]{huber_test}}
}
