% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mTest.R
\name{m_estimator_test}
\alias{m_estimator_test}
\title{Two sample location test based on M-estimators}
\usage{
m_estimator_test(x, y, alternative = c("two.sided", "greater", "less"),
  delta = 0, method = c("sampled", "exact"), psi = c("huber",
  "hampel", "bisquare"), k = .Mpsi.tuning.default(psi), n.rep = 10000,
  na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector of observations.}

\item{y}{a numeric vector of observations.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default),
"\code{greater}" or "\code{less}".}

\item{delta}{a numeric value indicating the true difference in the location parameter, default is \code{delta = 0}.}

\item{method}{a character string specifying the test method used: \code{"exact"} for an exact and \code{"sampled"}
for a randomized permutation test. The exact permutation test uses all data splits into two samples while the randomized test draws \code{n.rep} random splits
with replacement.}

\item{psi}{kernel used for optimization, must be one of \code{"bisquare"}, \code{"hampel"} and \code{"huber"},
defaults to \code{"huber"}.}

\item{k}{tuning parameter(s) for the respective kernel function, defaults to parameters implemented in .Mpsi.tuning.default(psi) from
\code{robustbase} package.}

\item{n.rep}{an integer value specifying the number of random permutations used to calculate
the permutation distribution if \code{method = "sampled"},
ignored if \code{method = "exact"}. Default is 10000.}

\item{na.rm}{a logical value indicating whether NA values in \code{x} and \code{y} should be stripped before the computation proceeds.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the estimated difference in means.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
\code{m_estimator_test} performs a two-sample permutation or randomization test
             based on M-estimators of location and respective variances
}
\details{
Details on the tests implemented here can be found in Aboukalam (1992).
}
\examples{
x <- rnorm(20); y <- rnorm(20)
m_estimator_test(x, y, psi = "huber", method = "sampled", n.rep = 1000)
m_estimator_test(x, y, psi = "bisquare", method = "sampled", n.rep = 1000)
}
\references{
\insertRef{Abo92robu}{robTests}
}
