% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyMTest_experimental.R
\name{m_test}
\alias{m_test}
\title{Two sample location test based on M-estimators}
\usage{
m_test(
  x,
  y,
  alternative = c("two.sided", "greater", "less"),
  delta = ifelse(var.test, 1, 0),
  method = c("asymptotic", "permutation", "randomization"),
  psi = c("huber", "hampel", "bisquare"),
  k = robustbase::.Mpsi.tuning.default(psi),
  n.rep = 10000,
  na.rm = FALSE,
  var.test = FALSE,
  wobble = FALSE,
  wobble.seed = NULL
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{delta}{a numeric value indicating the true difference in the location or scale parameter, depending
on whether the test should be performed for a difference in location or in scale, default is
\code{delta = 0} for location test and \code{delta = 1} for scale test.}

\item{method}{a character string specifying the test method used,
\code{"asymptotic"} for an asymptotic test based on the
normal distribution, \code{"permutation"} for a permutation and
\code{"randomization"} for a randomization test.
The permutation test uses all data splits into two samples,
while the randomization test draws \code{n.rep} random splits
with replacement.}

\item{psi}{kernel used for optimization.
Must be one of \code{"bisquare"}, \code{"hampel"} and \code{"huber"};
defaults to \code{"huber"}.}

\item{k}{tuning parameter(s) for the respective kernel function,
defaults to parameters implemented in \code{\link[robustbase:lmrob.control]{.Mpsi.tuning.default(psi)}} in the package \pkg{robustbase}.}

\item{n.rep}{an integer value specifying the number of random splits used to
calculate the randomization distribution if \code{method = "randomization"},
ignored if \code{method = "permutation"} or \code{method = "asymptotic"};
default is \code{n.rep = 10000}.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}

\item{var.test}{logical value testing whether the samples should be compared for a difference in scale.}

\item{wobble.seed}{logical value indicating whether uniform noise should be added to the sample before
proceeding with the test, in order to make sure there are no duplicate values in the
sample that can cause the scale estimate to be 0. Only necessary for the permutation and
randomization version of the test. Defaults to FALSE.}

\item{seed}{Optional, numerical value used as a seed to be used for 'wobbling' of the samples.
If no seed is given, it is set randomly. In any case, the seed used will be returned in a
warning message. The argument is ignored if \code{wobble = FALSE}.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{parameter}{the degrees of freedom for the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the Huber M-estimates of \code{x} and \code{y}.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
\code{m_estimator_test} performs a two-sample location test
             based on M-estimators.
}
\details{
The test statistic is the difference of M-estimates for both samples
divided by a pooled tau-estimate for the within-sample variance.

When computing a randomization distribution based on randomly drawn splits
with replacement, the results of Smyth & Phipson (2010) to calculate the p-value
are used. The psi and rho functions, which are needed to obtain the M-estimates,
are computed via the implementations in the package \code{\link[robustbase]{robustbase}}.
The tau scale estimate is computed with the default parameter settings
of the function \code{\link[robustbase]{scaleTau2}}.

The test statistic is given by

TODO: Include test statistic

which follows asymptotically a standard normal distribution. However, this is
only guaranteed for normally distributed data. In case of a non-normal
distribution, the test might not keep the desired significance level. This
can be corrected by correcting the test statistic by a factor which has to
be determined individually for each distribution.

For \code{var.test = TRUE}, the test compares the two samples for a difference in scale.
This is achieved by log-transforming the original observations so that a potential
scale difference appears as a location difference between the transformed samples;
see Fried (2012). The sample cannot contain zeros due to the necessary log-transformation.
If it contains zeros, uniform noise is added to all variables in order to remove zeros.
A warning is printed.

If the sample has been modified (either because of 0's for \code{var.test = TRUE}, or
as \code{wobble = TRUE}, the modified samples can be retrieved using

\code{set.seed(wobble.seed); wobble(x, y)}
}
