% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{mest_perm_distribution}
\alias{mest_perm_distribution}
\title{Permutation distribution for M-statistics}
\usage{
mest_perm_distribution(x, y, psi, k1, sampled = FALSE, n.rep = NULL)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{psi}{kernel used for optimization.
Must be one of \code{"bisquare"}, \code{"hampel"} and \code{"huber"};
defaults to \code{"huber"}.}

\item{sampled}{logical value indicating whether all splits should be
considered or only a random sample, where the observations
are drawn with replacement.}

\item{n.rep}{an integer value specifying the number of random permutations used to calculate
the permutation distribution if \code{method = "sampled"},
ignored if \code{method = "exact"}. Default is 10000.}

\item{k}{tuning parameter(s) for the respective kernel function,
defaults to parameters implemented in \code{\link[robustbase:lmrob.control]{.Mpsi.tuning.default(psi)}} in the package \pkg{robustbase}.}
}
\value{
Vector with permutation distribution.
}
\description{
\code{mest_perm_distribution} calculates the permutation distribution for M-statistics from
             \code{m_test_statistic}.
}
