% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data for the robust two sample tests}
\usage{
preprocess_data(x, y, delta, na.rm, wobble, wobble.seed, var.test)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{delta}{a numeric value indicating the true difference in the location or scale parameter, depending
on whether the test should be performed for a difference in location or in scale.
The default is \code{delta = 0} for location test and \code{delta = 1} for scale test.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}

\item{wobble}{logical value indicating whether uniform noise should be added to the sample before
proceeding with the test, in order to make sure there are no duplicate values in the
sample that can cause the scale estimate to be 0. Only necessary for the permutation and
randomization version of the test. The default is \code{wobble = FALSE}.}

\item{wobble.seed}{integer value used as a seed for the random number
generation in case of \code{wobble = TRUE} or when the
\code{var.test = TRUE} and on of the vectors \code{x} and
\code{y} contains zeros. When no seed is specified, it is
set randomly. The argument is ignored if
\code{var.test = FALSE} and \code{wobble = FALSE}.}

\item{var.test}{logical value testing whether the samples should be compared for a difference in scale.}
}
\value{
A named list containing the following components:
        \item{\code{x}}{the (possibly transformed) input vector \code{x}.}
        \item{\code{y}}{the (possibly transformed) input vector \code{y}.}
        \item{\code{delta}}{the (possibly transformed) input value
                            \code{delta}.}
}
\description{
\code{preprocess_data} is a helper function that performs several
preprocessing steps on the data before performing the two-sample tests.
}
\details{
The preprocessing steps include the removal of missing values and, if
specified, wobbling and a transformation of the observations to test for
differences in scale.
}
\keyword{internal}
