% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_missing_values}
\alias{remove_missing_values}
\title{Remove missing values in \code{x} and \code{y}}
\usage{
remove_missing_values(x, y, na.rm)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}
}
\value{
If one of the sample contains a missing value and \code{na.rm = FALSE},
\code{NA_real_} is returned. If the number of non-missing observations
in one of the samples is less than five, the function stops with an error
message. Otherwise, the function returns a named list
containing the following compoents:
\item{\code{x}}{input vector \code{x} without missing values.}
\item{\code{y}}{input vector \code{y} without missing values.}
}
\description{
\code{remove_missing_values} is a helper function that removes missing values
from the input sample vectors.
}
\details{
Both samples need at least five non-NA observations.
}
\keyword{internal}
