% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleEstimators.R
\name{win_var}
\alias{win_var}
\title{Winsorized variance}
\usage{
win_var(x, gamma = 0, na.rm = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{gamma}{numeric value in [0, 0.5] specifying the fraction of observations to be replaced
at each end of the sample before calculating the mean.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}
}
\value{
A list containing the following items:
\item{var}{The winsorized variance, and}
\item{h}{the degrees of freedom used for tests based on trimmed means and the
winsorized variance.}
}
\description{
\code{win_var} calculates the winsorized variance of a sample.
}
\examples{
set.seed(108)
x <- rnorm(10)
win_var(x, gamma = 0.2)


}
