% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobble.R
\name{wobble}
\alias{wobble}
\title{Test}
\usage{
wobble(x, y, partially = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{partially}{logical, indicates whether the complete sample should be treated or only the
values that are duplicated, default is \code{FALSE}}
}
\value{
A list of length two containing the modified \code{x} and \code{y}
}
\description{
\code{wobble} makes a discrete sample continuous by adding uniform noise to the
discrete observations
}
\details{
Test
}
\examples{
x <- rnorm(20); y <- rnorm(20); x <- round(x)
wobble(x, y)
wobble(x, y, partially=TRUE)

}
\references{
\insertRef{Fri12onli}{robTests}
}
