% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{asym_trimmed_perm_distribution}
\alias{asym_trimmed_perm_distribution}
\title{Permutation distribution for asymmetrically trimmed statistics}
\usage{
asym_trimmed_perm_distribution(x, y, type, sampled = FALSE,
  n.rep = NULL)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{type}{specifies the skewness selector statistic used for trimming, must be in
\code{"Q2", "SK2"} and \code{"SK5"}. Default is \code{"Q2"}.}

\item{sampled}{logical value indicating if the exact permutation distribution (\code{FALSE})
or only \code{n.rep} samples drawn with replacement (\code{TRUE})
should be used. Default is FALSE.}

\item{n.rep}{integer specifying the sample size in case of taking a random
sample to approximate the permutation distribution.}
}
\value{
Vector with permutation distribution.
}
\description{
\code{asym_trimmed_perm_distribution} calculates the permutation distribution for the asymmetrically
trimmed statistics from \code{asym_trimmed_test}.
}
