% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{asym_trimmed_perm_distribution}
\alias{asym_trimmed_perm_distribution}
\title{Permutation distribution for asymmetrically trimmed statistics}
\usage{
asym_trimmed_perm_distribution(x, y, type, sampled = FALSE, n.rep = NULL)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{type}{specifies the skewness selector statistic used for trimming, must be in
\code{"Q2", "SK2"} and \code{"SK5"}. Default is \code{"Q2"}.}

\item{sampled}{logical value indicating whether all splits should be
considered or only a random sample, where the observations
are drawn with replacement.}

\item{n.rep}{an integer value specifying the number of random splits used to
calculate the permutation distribution if \code{method = "sampled"},
ignored if \code{method = "exact"} or \code{method = "asymptotic"};
default is \code{n.rep = 10000}.}
}
\value{
Vector with permutation distribution.
}
\description{
\code{asym_trimmed_perm_distribution} calculates the permutation distribution for the asymmetrically
trimmed statistics from \code{asym_trimmed_test}.
}
