% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{asym_trimmed_t}
\alias{asym_trimmed_t}
\title{Test statistic for the asymmetrically trimmed test}
\usage{
asym_trimmed_t(x, y, type, delta = 0, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{y}{numeric vector of observations.}

\item{type}{specifies the skewness selector statistics, must be in \code{"SK2", "SK5", "Q2"}.}

\item{delta}{numeric indicating the true difference in means, default is \code{delta = 0}.}

\item{na.rm}{a logical value indicating whether NA values in \code{x} should be stripped before the computation proceeds.}
}
\value{
A list containing the following components:
\item{statistic}{the value of the test statistic.}
\item{estimates}{the asymmetrically trimmed means for both samples.}
\item{df}{the degrees of freedom for the test statistic.}
}
\description{
\code{asym_trimmed_t} calculates the test statistic of the asymmetrically trimmed test with a given skewness selector statistic
}
\references{
\insertRef{YueDix73appr}{robTests}

\insertRef{Yue74trim}{robTests}
}
