% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymTrimmedTest.R
\name{asym_trimmed_test}
\alias{asym_trimmed_test}
\title{Two-sample location tests based on asymmetrically trimmed means}
\usage{
asym_trimmed_test(x, y, type = c("Q2", "SK2", "SK5"),
  alternative = c("two.sided", "greater", "less"), delta = 0,
  method = c("asymptotic", "exact", "sampled"), n.rep = 10000,
  na.rm = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{type}{specifies the skewness selector statistic used for trimming, must be in
\code{"Q2", "SK2"} and \code{"SK5"}. Default is \code{"Q2"}.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{delta}{a numeric value indicating the true difference in the location parameter, default is delta = 0.}

\item{method}{a character string specifying the test method used, "asymptotic" for an asymptotic test based on the normal distribution, "exact" for an exact and "sampled" for a randomized permutation test. The exact permutation test uses all data splits into two samples while the randomized test draws n.rep random splits with replacement.}

\item{n.rep}{an integer value specifying the number of random splits used to calculate the permutation distribution if method = "sampled", ignored if method = "exact" or method = "asymptotic". Default is n.rep = 10000.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the estimated sample means of \code{x} and \code{y}.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
\code{asym_trimmed_test()} performs a two-sample location test by using
asymmetrically trimmed means based on different skewness-
selector statistics for both samples.
}
\details{
The test statistic is an analogue to the one of the ordinary t-test, where
the difference of the sample means is replaced by the difference of asymmetrically trimmed means
and the pooled empirical standard deviation is replaced by a pooled winsorized standard deviation
which uses the asymmetrically trimmed means.

The number of observations trimmed from the lower and the upper end of the sample
depends on the skewness. Reed & Stark (2004) suggest three possible skewness-selector statistics.
The argument \code{type} specifies which one is used.

\describe{
\item{"Q2"}{This skewness-selector statistic compares the difference between the upper and the lower 5\% of
            the observations to the difference between the upper and lower 50\%.}
\item{"SK2"}{This skewness-selector statistic compares the difference between the sample minimum and
             median to the difference between the median and the maximum.}
\item{"SK5"}{This skewness-selector statistic compares the difference between the sample minimum and
             mean to the difference between the mean and the maximum.}
}
}
\references{
\insertRef{ReeSta04robu}{robTests}
}
