% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{calc_perm_p_value}
\alias{calc_perm_p_value}
\title{Calculation of permutation p-value}
\usage{
calc_perm_p_value(statistic, distribution, m, n, sampled, n.rep,
  alternative)
}
\arguments{
\item{statistic}{observed value of the test statistic.}

\item{distribution}{numeric vector with permutation distribution.}

\item{m}{integer value giving size of first sample.}

\item{n}{integer value giving size of second sample.}

\item{sampled}{logical value specifying whether the distribution consists of
all permutations or only a random subset drawn with replacement.}

\item{n.rep}{integer value giving the number of replications
(only needed if \code{sampled = TRUE}).}

\item{alternative}{character string specifying the alternative hypothesis,
must be one of "\code{two.sided}" (default),
"\code{greater}" or "\code{less}".}
}
\value{
p.value for the specified alternative.
}
\description{
\code{calc_perm_p_value} calculates the permutation p-value following Smyth & Phipson (2010).
}
\references{
\insertRef{SmyPhi10perm}{robTests}
}
