% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl1Test.R
\name{hl1_test}
\alias{hl1_test}
\title{Two-sample location tests based on one-sample Hodges-Lehmann estimator}
\usage{
hl1_test(
  x,
  y,
  alternative = c("two.sided", "greater", "less"),
  delta = ifelse(var.test, 1, 0),
  method = c("asymptotic", "permutation", "randomization"),
  scale = c("S1", "S2"),
  n.rep = 10000,
  na.rm = FALSE,
  var.test = FALSE
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{delta}{a numeric value indicating the true difference in the location or scale parameter, depending
on whether the test should be performed for a difference in location or in scale, default is
\code{delta = 0} for location test and \code{delta = 1} for scale test.}

\item{method}{a character string specifying the test method used,
\code{"asymptotic"} for an asymptotic test based on the
normal distribution, \code{"permutation"} for a permutation and
\code{"randomization"} for a randomization test.
The permutation test uses all data splits into two samples,
while the randomization test draws \code{n.rep} random splits
with replacement.}

\item{scale}{a character string specifying the scale estimator used for standardization
of the test statistic, must be one of \code{"S1"} and \code{"S2"}.
Default is \code{"S1"}. Ignored if \code{method = "asymptotic"};
see details for definition of the scale estimators.}

\item{n.rep}{an integer value specifying the number of random splits used to
calculate the randomization distribution if \code{method = "randomization"},
ignored if \code{method = "permutation"} or \code{method = "asymptotic"};
default is \code{n.rep = 10000}.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}

\item{var.test}{logical value testing whether the samples should be compared for a difference in scale.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the one-sample Hodges-Lehmann estimates of \code{x} and \code{y}.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
\code{hl1_test} performs a two-sample location test based on
the difference of the one-sample Hodges-Lehmann estimators for both samples.
}
\details{
When computing a randomization distribution based on randomly drawn splits with replacement, the results of
Smyth & Phipson (2010) to calculate the p-value are used. The test statistics and the asymptotic distribution are taken from Fried & Dehling (2011).

The test statistics for the permutation and randomization version of the test is standardized using a robust scale estimator.
\code{scale = "S1"} represents use of

\deqn{S = med(|X_i - X_j|: 1 \le i < j \le m, |Y_i - Y_j|, 1 <= i < j <= n),}

\code{scale = "S2"} uses

\deqn{S = med(|Z_i - Z_j|: 1 \le i < j \le m+n) }

where \eqn{ Z = ( X_1 - med(X),...,X_m - med(X), Y_1 - med(Y),...,Y_n - med(Y) )'}
is the median corrected sample. For more details see Fried & Dehling (2011).

For \code{var.test = TRUE}, the test compares two sample for a difference in scale.
This is achieved by log-transforming the original observations so that a potential
scale difference appears as a location difference between the transformed samples;
see Fried (2012).
}
\examples{
## Generate random samples
set.seed(108)
x <- rnorm(20); y <- rnorm(20)

## Asymptotic HL1 test
hl1_test(x, y, method = "asymptotic", scale = "S1")

## HL12 test using randomization principle by drawing 1000 random permutations
## with replacement
\dontrun{
hl1_test(x, y, method = "randomization", n.rep = 1000, scale = "S2")
}

}
\references{
\insertRef{PhiSmy10perm}{robTests}

\insertRef{FriDeh11robu}{robTests}

\insertRef{Fri12onli}{robTests}
}
