% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationEstimators.R
\name{hodges_lehmann}
\alias{hodges_lehmann}
\title{One-sample Hodges-Lehmann estimator}
\usage{
hodges_lehmann(x, na.rm = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{na.rm}{a logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}
}
\value{
The one-sample Hodges-Lehmann estimator.
}
\description{
\code{hodges_lehmann} calculates the one-sample Hodges-Lehmann estimator
of a sample.
}
\examples{
set.seed(108)
x <- rnorm(10)
hodges_lehmann(x)

}
\references{
\insertRef{HodLeh63esti}{robTests}
}
