% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberTest.R
\name{huber_test}
\alias{huber_test}
\title{Two-sample location tests based on Huber's M-estimator}
\usage{
huber_test(
  x,
  y,
  delta = 0,
  k = 1.8,
  alternative = c("two.sided", "greater", "less"),
  var.test = FALSE
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{delta}{a numeric value indicating the true difference in the location or scale parameter, depending
on whether the test should be performed for a difference in location or in scale, default is
\code{delta = 0} for location test and \code{delta = 1} for scale test.}

\item{k}{tuning parameter for Huber's M-estimator, default is \code{k = 1.8}.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{var.test}{logical value testing whether the samples should be compared for a difference in scale.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{parameter}{the degrees of freedom for the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the Huber M-estimates of \code{x} and \code{y}.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of test was performed.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
Performs a 2-sample location test based on Huber's M-estimator
}
\details{
The test is introduced in the context of the hybrid tests in Weichert & Hothorn (2002).
}
\examples{
x <- rnorm(20); y <- rnorm(20)
huber_test(x, y)

}
\references{
\insertRef{WeiHot02robu}{robTests}
}
\seealso{
\code{\link[robTests]{min_tc_test}}
\code{\link[robTests]{min_t_test}}
}
