% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationEstimators.R
\name{m_est}
\alias{m_est}
\title{M-estimator of location}
\usage{
m_est(x, psi, k = .Mpsi.tuning.default(psi), tol = 1e-06, max.it = 15)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{psi}{kernel used for optimization, must be one of \code{"bisquare"}, \code{"hampel"} and \code{"huber"}}

\item{k}{tuning parameter(s) for the respective kernel function, defaults to parameters implemented in .Mpsi.tuning.default(psi) from 
\code{robustbase} package.}

\item{tol}{tolerance for convergence, defaults to 1e-06.}

\item{max.it}{the maximum number of iterations, defaults to 15.}
}
\value{
A list containing the items:
        \item{est}{estimated mean, and}
        \item{var}{estimated variance.}
}
\description{
\code{m_est} calculates the M-estimate of location and its variance using different tuning functions
}
\examples{

}
