% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{m_test_statistic}
\alias{m_test_statistic}
\title{Calculate the M-test-statistic}
\usage{
m_test_statistic(x, y, psi, k = .Mpsi.tuning.default(psi))
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{k}{tuning parameter(s) for the respective kernel function,
defaults to parameters implemented in \code{\link[robustbase:lmrob.control]{.Mpsi.tuning.default(psi)}} in the package \pkg{robustbase}.}
}
\value{
A list containing
        \item{statistic}{the standardized test statistic, and}
        \item{estimates}{the M-estimates of location for both \code{x} and \code{y}}
}
\description{
\code{m_test_statistic} calculates the test statistics for randomization tests based on M-estimators
}
