% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{m_test_statistic}
\alias{m_test_statistic}
\title{Calculate the M-test-statistic}
\usage{
m_test_statistic(x, y, psi, k = .Mpsi.tuning.default(psi))
}
\arguments{
\item{x}{numeric vector of observations.}

\item{y}{numeric vector of observations.}

\item{psi}{kernel used for optimization, must be one of \code{"bisquare"}, \code{"hampel"} and \code{"huber"}}

\item{k}{tuning parameter(s) for the respective kernel function, defaults to parameters implemented in .Mpsi.tuning.default(psi) from 
\code{robustbase} package.}
}
\value{
A list containing
        \item{statistic}{the standardized test statistic, and}
        \item{estimates}{the M-estimates of location for both \code{x} and \code{y}}
}
\description{
\code{m_test_statistic} calculates the test statistics for randomization tests based on M-estimators
}
