% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{perm_distribution}
\alias{perm_distribution}
\title{Permutation Distribution}
\usage{
perm_distribution(x, y, type, randomization = FALSE, n.rep = 10000)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{type}{desired test statistic, in \code{"D1S1"} (default), \code{"D2S1", "D2S1", "D2S2", "D3S3", "D3S4"},
where "D1" to "D3" mark different estimators for the location shift and
"S1" to "S4" are estimates for the scale of the samples; defaults to \code{"D1S1"}.}

\item{randomization}{logical value indicating whether the p-value should be computed from a permutation (\code{FALSE}) or a randomization (\code{TRUE}) distribution.}

\item{n.rep}{an integer value specifying the number of random splits used to
calculate the randomization distribution if \code{method = "randomization"},
ignored if \code{method = "permutation"} or \code{method = "asymptotic"};
default is \code{n.rep = 10000}.}
}
\value{
Vector with permutation distribution.
}
\description{
\code{perm_distribution()} calculates the permutation distribution for
several test statistics.
}
\details{
see documentation of \code{rob_perm_statistic()} for a description of
the \code{type}-parameter
}
