% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDistribution.R
\name{perm_distribution}
\alias{perm_distribution}
\title{Permutation Distribution}
\usage{
perm_distribution(x, y, type, sampled = FALSE, n.rep = NULL)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{y}{numeric vector of observations.}

\item{type}{desired test statistic, in \code{"D1S1"} (default), \code{"D2S1", "D2S1", "D2S2", "D3S3", "D3S4"},
where "D1" to "D3" mark different estimators for the location shift and
"S1" to "S4" are estimates for the scale of the samples; defaults to \code{"D1S1"}.}

\item{sampled}{logical value indicating whether all permutations should be
considered or only a random sample.}

\item{n.rep}{integer specifying the sample size in case of taking a random
sample to approximate the permutation distribution.}
}
\value{
Vector with permutation distribution.
}
\description{
\code{perm_distribution()} calculates the permutation distribution for
several test statistics.
}
\details{
see documentation of \code{rob_perm_statistic()} for a description of
the \code{type}-parameter
}
