% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{rob_perm_statistic}
\alias{rob_perm_statistic}
\title{Robust permutation statistics based on medians}
\usage{
rob_perm_statistic(x, y, type = c("D1S1", "D1S2", "D2S1", "D2S2", "D3S3",
  "D3S4"), na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{y}{numeric vector of observations.}

\item{type}{desired test statistic, in \code{"D1S1"} (default), \code{"D2S1", "D2S1", "D2S2", "D3S3", "D3S4"},
where "D1" to "D3" mark different estimators for the location shift and
"S1" to "S4" are estimates for the scale of the samples; defaults to \code{"D1S1"}.}

\item{na.rm}{a logical value indicating whether NA values in \code{x} should be stripped before the computation proceeds;
defaults to \code{FALSE}.}
}
\value{
A list containing the following components:
        \item{statistic}{the selected test statistic}
        \item{abs.statistic}{the absolute value of the selected test statistic}
        \item{estimates}{estimate of location for each sample if available}
}
\description{
\code{rob_perm_statistic()} calculates test statistics for robust permutation tests based on medians
}
\details{
The test statistics returned by \code{rob_perm_statistic} are of the form
         \deqn{D_i/S_j} where the D_i, i = 1,...,3, are different estimators of location and the S_j, j = 1,...,4 are
         estimates for the mutual sample scale. See Fried and Dehling (2011) for details.
}
\examples{
x <- rnorm(20); y <- rnorm(20)
rob_perm_statistic(x, y, type = "D2S1")

}
\references{
\insertRef{FriDeh11robu}{robTests}
}
