% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_method}
\alias{select_method}
\title{Select principle for computing null distribution}
\usage{
select_method(x, y, method, test.name)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{test.name}{character string specifying the two-sample test for which
the helper function is used.}
}
\value{
A character string which contains the principle for computing the
        null distribution.
}
\description{
\code{select_method} is a helper function that chooses the principle for
computing the null distribution of a two-sample test.
}
\details{
When the principle is specified by the user, i.e. method contains only
one element, the selected method is returned. Otherwisse, if the user
does not specify the principle, it depends on the sample size: When both
samples contain more than 30 observations, an asymptotic test is perfomed.
If one of the samples contains less than 30 observations, the null
distribution is computed via the randomization principle. The number of
replications \code{n.rep} for the randomization test needs to be specified
outside of this function. Each test function contains the argument
\code{n.rep} where this can be done.
}
\keyword{internal}
