% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationEstimators.R
\name{trim_mean}
\alias{trim_mean}
\title{Trimmed mean}
\usage{
trim_mean(x, gamma = 0.2, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{gamma}{numeric value in [0, 0.5] specifying the fraction of observations to be trimmed 
from each end of the sample before calculating the mean.}

\item{na.rm}{a logical value indicating whether NA values in \code{x} should be stripped before the computation proceeds.}
}
\value{
The trimmed mean.
}
\description{
\code{trim_mean} calculates a trimmed mean of a sample.
}
\details{

}
\examples{
x <- rnorm(10)
trim_mean(x, gamma = 0.2)

}
\references{
\insertRef{ReeSta96hing}{robTests}
}
