% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{trimmed_t}
\alias{trimmed_t}
\title{Test statistic for the two-sample Yuen t-test}
\usage{
trimmed_t(x, y, gamma = 0.2, delta = 0, na.rm = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{gamma}{numeric value in [0, 0.5] specifying the fraction of observations to be
trimmed from each end of the sample before calculating the mean. Values
of trim outside that range are taken as the nearest endpoint.}

\item{delta}{a numeric value indicating the true difference in the location or scale parameter, depending
on whether the test should be performed for a difference in location or in scale.
The default is \code{delta = 0} for location test and \code{delta = 1} for scale test.}

\item{na.rm}{logical value indicating whether NA values in x and y should be stripped before the computation proceeds.}
}
\value{
A list containing the following components:
\item{statistic}{the value of the test statistic.}
\item{estimates}{the trimmed means for both samples.}
\item{df}{the degrees of freedom for the test statistic.}
}
\description{
\code{trimmed_t} calculates the test statistic of the two-sample Yuen test.
}
\references{
\insertRef{YueDix73appr}{robTests}

\insertRef{Yue74trim}{robTests}
}
