% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{trimmed_t}
\alias{trimmed_t}
\title{Test statistic for the two-sample Yuen t-test}
\usage{
trimmed_t(x, y, gamma = 0.2, delta = 0, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{y}{numeric vector of observations.}

\item{gamma}{numeric value in [0, 0.5] specifying the fraction of observations to be trimmed from each end of the sample before calculating the mean. Values of trim outside that range are taken as the nearest endpoint.}

\item{delta}{numeric indicating the true difference in means, default is \code{delta = 0}.}

\item{na.rm}{a logical value indicating whether NA values in \code{x} should be stripped before the computation proceeds.}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}
}
\value{
A list containing the following components:
\item{statistic}{the value of the test statistic.}
\item{estimates}{the trimmed means for both samples.}
\item{df}{the degrees of freedom for the test statistic.}
}
\description{
\code{trimmed_t} calculates the test statistic of the two-sample Yuen test.
}
\references{
\insertRef{YueDix73appr}{robTests}

\insertRef{Yue74trim}{robTests}
}
