% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmedTest.R
\name{trimmed_test}
\alias{trimmed_test}
\title{Two-sample Trimmed t-test (Yuen's t-Test)}
\usage{
trimmed_test(x, y, gamma = 0.2, alternative = c("two.sided", "less",
  "greater"), delta = 0, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of observations.}

\item{y}{numeric vector of observations.}

\item{gamma}{numeric value in [0, 0.5] specifying the fraction of observations to be trimmed from each end of the sample before calculating the mean. Values of trim outside that range are taken as the nearest endpoint.}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}

\item{delta}{numeric indicating the true difference in means}

\item{na.rm}{a logical value indicating whether NA values in \code{x} should be stripped before the computation proceeds.}
}
\value{
A list with class "\code{htest}" containing the following components:
\item{statistic}{the value of the test statistic.}
\item{parameter}{the degrees of freedom for the test statistic.}
\item{p.value}{the p-value for the test.}
\item{estimate}{the estimated difference in means.}
\item{null.value}{the specified hypothesized value of the mean difference.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
\code{trimmed_test} performs the two-sample Yuen t-test.
}
\examples{
x <- rnorm(20); y <- rnorm(20)
trimmed_test(x, y, gamma = 0.1)

}
\references{
\insertRef{YueDix73appr}{robTests}

\insertRef{Yue74trim}{robTests}
}
