% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobble.R
\name{wobble}
\alias{wobble}
\title{Funtion that undiscretizes rounded samples}
\usage{
wobble(x, y, check = TRUE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{check}{logical value indicating whether the samples should be checked for bindings prior to
adding uniform noise or not, defaults to \code{TRUE}}
}
\value{
A list of length two containing the modified \code{x} and \code{y}
}
\description{
\code{wobble} makes a discrete sample with duplicated values continuous by adding uniform noise to the
discrete observations
}
\details{
If \code{check = TRUE} the function checks whether all values in the two numeric input vectors are distinct.
If so, it returns the original values, otherwise the values are made continuous by adding uniform
noise. Otherwise, it simply determines the number of digits and adds uniform noise in order to
make the sample "more" continuous.
}
\examples{
x <- rnorm(20); y <- rnorm(20); x <- round(x)
wobble(x, y)

}
\references{
\insertRef{Fri12onli}{robTests}
}
